refactor(tui): extract the pane grid and chrome into tui::render::panes
The brand and tab strip, per-pane titles and borders, and both
multi-pane entry points move out of the module root with the sixteen
tests that render them.
