+++ src/tui.rs
@@ -27,7 +27,6 @@ use std::{
     time::{Duration, Instant},
 };
 use tokio::sync::{mpsc, watch};
-use unicode_width::UnicodeWidthStr;
 
 use serde::{Deserialize, Serialize};
 
@@ -45,14 +44,7 @@ use crossterm::{
 };
 use iroh::EndpointAddr;
 use portable_pty::PtySize;
-use ratatui::{
-    Frame, Terminal, TerminalOptions, Viewport,
-    backend::CrosstermBackend,
-    layout::{Alignment, Rect},
-    style::{Color, Modifier, Style},
-    text::{Line, Span},
-    widgets::{Block, Clear, Paragraph},
-};
+use ratatui::{Terminal, TerminalOptions, Viewport, backend::CrosstermBackend, layout::Rect};
 
 use crate::{
     agent_detect::{
@@ -102,16 +94,16 @@ use input::{
     },
     mouse::encode_mouse,
 };
-use render::agents::{AGENT_OVERLAY_CARD_LINES, agents_overlay_content, render_agents_overlay};
-use render::footer::render_contextual_footer;
-use render::modals::{render_delete_tab_confirmation, render_rename_prompt, render_share_modal};
+use render::agents::{AGENT_OVERLAY_CARD_LINES, agents_overlay_content};
+use render::panes::{
+    TAB_BAR_SEPARATOR, TOP_BAR_BRAND, TOP_BAR_BRAND_SEPARATOR, render_shared_multi_pane, tab_label,
+};
+pub use render::panes::{render_multi_pane, render_multi_pane_with_copy_feedback};
 use render::vt::{
     VtScreen, available_scrollback, render_guest_screen, render_host_screen, viewed_screen,
 };
 pub(crate) use share::{resolve_local_ticket, share_copy_result};
-use text::{
-    copied_line_count, sanitize_single_line, text_width, truncate_bytes, truncate_trailing,
-};
+use text::{copied_line_count, sanitize_single_line, text_width, truncate_bytes};
 
 pub(crate) enum NodeScreenSnapshot {
     Local {
@@ -136,9 +128,6 @@ pub(crate) struct LocalScrollbackWindow {
 /// Kept as the module's public marker from the scaffold.
 pub struct Tui;
 
-const TOP_BAR_BRAND: &str = "p2pmux";
-const TOP_BAR_BRAND_SEPARATOR: &str = " │ ";
-const TAB_BAR_SEPARATOR: &str = " · ";
 /// The legacy fixed-grid host/guest footer, which has no chords, agents, or share modal.
 const CONTROL_HELP: &str = "Ctrl+ <p> PANE   <t> TAB   <q> QUIT   Option+ <shift> + <↑↓←→> FOCUS";
 const CHORD_IDLE_TIMEOUT: Duration = Duration::from_secs(2);
@@ -1853,75 +1842,6 @@ impl MultiPaneTui {
     }
 }
 
-fn tab_label(title: Option<&str>, index: usize, active: bool, unread: bool) -> String {
-    let label = title.map_or_else(|| format!("Tab #{index}"), str::to_owned);
-    let label = if unread { format!("* {label}") } else { label };
-    if active { format!(" {label} ") } else { label }
-}
-
-#[allow(clippy::too_many_arguments)]
-fn pane_title(
-    custom_title: Option<&str>,
-    index: usize,
-    host_peer_id: &[u8],
-    controller_peer_id: Option<&[u8]>,
-    locked: bool,
-    exited: bool,
-    members: &[crate::layout::Member],
-    available_width: usize,
-) -> Line<'static> {
-    let control = match controller_peer_id {
-        Some([]) => "free".to_owned(),
-        Some(peer_id) => member_label(peer_id, members),
-        None => "…".to_owned(),
-    };
-    let label = truncate_trailing(
-        &custom_title.map_or_else(|| format!("Pane #{index}"), str::to_owned),
-        available_width,
-    );
-    let control = if exited {
-        "exited"
-    } else if locked {
-        "host-only"
-    } else {
-        &control
-    };
-    let metadata = format!(
-        " host: {} control: {control}",
-        member_label(host_peer_id, members)
-    );
-    let metadata_width = available_width.saturating_sub(UnicodeWidthStr::width(label.as_str()));
-    Line::from(vec![
-        Span::styled(label, Style::default().add_modifier(Modifier::BOLD)),
-        Span::raw(truncate_trailing(&metadata, metadata_width)),
-    ])
-}
-
-fn pane_border_color(
-    theme: &UiTheme,
-    exited: bool,
-    controller_peer_id: Option<&[u8]>,
-    _controller_active: bool,
-    focused: bool,
-    hovered: bool,
-    chord_mode: ChordMode,
-) -> Color {
-    if exited {
-        return theme.pane_border_idle;
-    }
-    if focused && chord_mode == ChordMode::Pane {
-        return theme.pane_border_chord_focused;
-    }
-
-    match controller_peer_id {
-        Some([]) if focused => theme.pane_border_free_focused,
-        None if focused => theme.pane_border_unknown_focused,
-        Some([]) | None if hovered => theme.pane_border_hovered,
-        Some([]) | None => theme.pane_border_idle,
-        Some(_) => theme.pane_border_remote_control,
-    }
-}
-
 fn selection_text(screen: &vt100::Screen, selection: PaneTextSelection) -> Option<String> {
     if selection.is_empty() {
         return None;
@@ -1977,298 +1897,6 @@ fn copy_to_macos_clipboard(text: &str) -> io::Result<()> {
         Ok(())
     } else {
         Err(io::Error::other("pbcopy failed"))
-    }
-}
-
-/// Renders layout chrome plus any currently available fixed-size VT screens.
-pub fn render_multi_pane(
-    frame: &mut Frame<'_>,
-    tui: &MultiPaneTui,
-    screens: &BTreeMap<PaneId, &vt100::Screen>,
-) {
-    render_shared_multi_pane(
-        frame,
-        tui,
-        screens,
-        "",
-        None,
-        None,
-        ShareView::default(),
-        None,
-    );
-}
-
-/// Renders the local attachment footer with its own copy feedback.
-#[allow(clippy::too_many_arguments)]
-pub fn render_multi_pane_with_copy_feedback(
-    frame: &mut Frame<'_>,
-    tui: &MultiPaneTui,
-    screens: &BTreeMap<PaneId, &vt100::Screen>,
-    copied_lines: Option<usize>,
-    footer_notice: Option<&str>,
-    share: ShareView<'_>,
-    local_peer_id: Option<&[u8]>,
-    link: Option<&str>,
-) {
-    let exited_notice = tui
-        .snapshot()
-        .panes
-        .get(&tui.focused_pane())
-        .filter(|pane| pane.exited)
-        .map(|pane| {
-            if local_peer_id == Some(pane.host_peer_id.as_slice()) {
-                "exited — close with Ctrl+P, X"
-            } else {
-                "exited — input disabled; pane host can close with Ctrl+P, X"
-            }
-        });
-    render_shared_multi_pane(
-        frame,
-        tui,
-        screens,
-        "",
-        copied_lines,
-        footer_notice.or(exited_notice),
-        share,
-        link,
-    );
-}
-
-#[allow(clippy::too_many_arguments)]
-fn render_shared_multi_pane(
-    frame: &mut Frame<'_>,
-    tui: &MultiPaneTui,
-    screens: &BTreeMap<PaneId, &vt100::Screen>,
-    status: &str,
-    copied_lines: Option<usize>,
-    footer_notice: Option<&str>,
-    share: ShareView<'_>,
-    link: Option<&str>,
-) {
-    let theme = &tui.theme;
-    let geometry = tui.geometry(frame.area());
-    if geometry.tab_bar.width > 0 && geometry.tab_bar.height > 0 {
-        let buffer = frame.buffer_mut();
-        buffer.set_stringn(
-            geometry.tab_bar.x,
-            geometry.tab_bar.y,
-            " ".repeat(usize::from(geometry.tab_bar.width)),
-            usize::from(geometry.tab_bar.width),
-            Style::default().bg(theme.footer_background),
-        );
-        let mut x = buffer
-            .set_stringn(
-                geometry.tab_bar.x,
-                geometry.tab_bar.y,
-                tui.title(),
-                usize::from(geometry.tab_bar.width),
-                Style::default()
-                    .fg(theme.tab_foreground)
-                    .bg(theme.footer_background),
-            )
-            .0;
-        x = buffer
-            .set_stringn(
-                x,
-                geometry.tab_bar.y,
-                TOP_BAR_BRAND_SEPARATOR,
-                usize::from(geometry.tab_bar.right().saturating_sub(x)),
-                Style::default()
-                    .fg(theme.tab_separator)
-                    .bg(theme.footer_background),
-            )
-            .0;
-        for (index, tab) in tui.snapshot.tabs.iter().enumerate() {
-            if index > 0 {
-                x = buffer
-                    .set_stringn(
-                        x,
-                        geometry.tab_bar.y,
-                        TAB_BAR_SEPARATOR,
-                        usize::from(geometry.tab_bar.right().saturating_sub(x)),
-                        Style::default()
-                            .fg(theme.tab_separator)
-                            .bg(theme.footer_background),
-                    )
-                    .0;
-            }
-            let active = tab.tab_id == tui.current_tab;
-            let style = if active {
-                Style::default()
-                    .fg(theme.tab_foreground)
-                    .bg(theme.tab_active_background)
-                    .add_modifier(Modifier::BOLD)
-            } else {
-                Style::default()
-                    .fg(if tui.chord_mode == ChordMode::Tab {
-                        theme.tab_separator
-                    } else {
-                        theme.tab_foreground
-                    })
-                    .bg(theme.footer_background)
-            };
-            let label = truncate_trailing(
-                &tab_label(
-                    tab.title.as_deref(),
-                    index + 1,
-                    active,
-                    tui.tab_has_unread_agent_pane(tab),
-                ),
-                usize::from(geometry.tab_labels[&tab.tab_id].width),
-            );
-            let label_rect = geometry.tab_labels[&tab.tab_id];
-            if label_rect.width == 0 {
-                continue;
-            }
-            buffer.set_stringn(
-                label_rect.x,
-                label_rect.y,
-                &label,
-                usize::from(label_rect.width),
-                style,
-            );
-            x = x.saturating_add(text_width(&label));
-        }
-        // `direct 35ms` / `relayed 120ms ×3`, right-aligned. Lives in the tab bar rather
-        // than the footer because the footer is transient -- chords, copy feedback and
-        // status notices all take it over -- and connectivity is exactly the fact you
-        // want on screen at the moment something else has gone wrong.
-        let badge = match (tui.session_locked, link.filter(|link| !link.is_empty())) {
-            (true, Some(link)) => Some(format!("locked · {link}")),
-            (true, None) => Some(String::from("locked")),
-            (false, link) => link.map(str::to_owned),
-        };
-        if let Some(link) = badge.as_deref() {
-            let width = text_width(link);
-            let start = geometry.tab_bar.right().saturating_sub(width);
-            // Never overwrite a tab label: if the tabs already reach that far, the tab
-            // names are the more important thing and the badge simply stands down.
-            if width > 0 && start > x {
-                buffer.set_stringn(
-                    start,
-                    geometry.tab_bar.y,
-                    link,
-                    usize::from(width),
-                    Style::default()
-                        .fg(theme.tab_separator)
-                        .bg(theme.footer_background),
-                );
-            }
-        }
-    }
-    if geometry.footer.width > 0 && geometry.footer.height > 0 {
-        render_contextual_footer(
-            frame.buffer_mut(),
-            theme,
-            geometry.footer,
-            status,
-            copied_lines,
-            footer_notice,
-            tui.chord_mode,
-        );
-    }
-
-    let pane_ids = tui
-        .current_tab_layout()
-        .map(|tab| visible_leaf_panes(&tab.root))
-        .unwrap_or_default();
-    for (index, pane_id) in pane_ids.into_iter().enumerate() {
-        let Some(rect) = geometry.panes.get(&pane_id).copied() else {
-            continue;
-        };
-        let pane = &tui.snapshot.panes[&pane_id];
-        let view = tui.pane_views.get(&pane_id).cloned().unwrap_or_default();
-        let focused = pane_id == tui.focused_pane;
-        let title_width = usize::from(rect.width.saturating_sub(2));
-        let lock_badge = (!pane.exited && pane.locked).then(|| {
-            format!(
-                "(locked by {})",
-                member_label(&pane.host_peer_id, &tui.snapshot.members)
-            )
-        });
-        let badge_width = lock_badge
-            .as_deref()
-            .map_or(0, |badge| UnicodeWidthStr::width(badge).min(title_width));
-        let mut title = pane_title(
-            pane.title.as_deref(),
-            index + 1,
-            &pane.host_peer_id,
-            view.controller_peer_id.as_deref(),
-            pane.locked,
-            pane.exited,
-            &tui.snapshot.members,
-            title_width.saturating_sub(badge_width).saturating_sub(2),
-        );
-        title.spans.insert(
-            0,
-            Span::raw(if tui.unread_agent_panes.contains(&pane_id) {
-                " * "
-            } else {
-                " "
-            }),
-        );
-        title.spans.push(Span::raw(" "));
-        let border_color = pane_border_color(
-            theme,
-            pane.exited,
-            view.controller_peer_id.as_deref(),
-            view.controller_active,
-            focused,
-            tui.hovered_pane == Some(pane_id),
-            tui.chord_mode,
-        );
-        let mut block = Block::bordered()
-            .title(title)
-            .border_style(Style::default().fg(border_color));
-        if let Some(badge) = lock_badge {
-            block = block.title(
-                Line::from(truncate_trailing(&badge, badge_width)).alignment(Alignment::Right),
-            );
-        }
-        let content = pane_content_rect(rect);
-        frame.render_widget(block, rect);
-        // The fixed VT grid may be smaller than this pane after layout reflow. Clear the full
-        // interior before drawing it so letterbox margins cannot retain cells from an older pane.
-        frame.render_widget(Clear, content);
-        if let Some(screen) = screens.get(&pane_id) {
-            let viewport = fixed_grid_viewport(content, pane.grid_rows, pane.grid_cols);
-            let screen = viewed_screen(screen, tui.scrollback_offset(pane_id));
-            frame.render_widget(
-                VtScreen::new(&screen).with_selection(
-                    tui.selection()
-                        .filter(|selection| selection.pane_id == pane_id),
-                ),
-                viewport,
-            );
-            let (row, col) = screen.cursor_position();
-            if focused
-                && view.ready
-                && screen.scrollback() == 0
-                && !screen.hide_cursor()
-                && !pane.exited
-                && row < viewport.height
-                && col < viewport.width
-            {
-                frame.set_cursor_position((
-                    viewport.x.saturating_add(col),
-                    viewport.y.saturating_add(row),
-                ));
-            }
-        } else if !view.ready {
-            frame.render_widget(Paragraph::new("waiting for pane snapshot/lease"), content);
-        }
-    }
-    if tui.overlay_open() {
-        render_agents_overlay(frame, tui, unix_ms_now());
-    }
-    if tui.share_open() {
-        render_share_modal(frame, &tui.theme, share);
-    }
-    if let ModalState::Rename(prompt) = &tui.modal {
-        render_rename_prompt(frame, prompt, &tui.theme);
-    }
-    if let ModalState::ConfirmDeleteTab { pane_count, .. } = &tui.modal {
-        render_delete_tab_confirmation(frame, *pane_count, &tui.theme);
     }
 }
 
@@ -5303,12 +4931,7 @@ mod tests {
 
     use crossterm::event::{KeyCode, KeyEvent, KeyModifiers, MouseButton, MouseEventKind};
     use portable_pty::PtySize;
-    use ratatui::{
-        Terminal,
-        backend::TestBackend,
-        layout::Rect,
-        style::{Color, Modifier},
-    };
+    use ratatui::{Terminal, backend::TestBackend, layout::Rect};
 
     use crate::layout::{Axis, LayoutError, NewPanePosition, Node, Tab};
     use crate::lease::{IDLE_AFTER, LeaseManager, LeaseState};
@@ -5327,12 +4950,10 @@ mod tests {
         AGENT_TOGGLE_WINDOW, CHORD_IDLE_TIMEOUT, ChordMode, HostControlEvent, HostPaneChannels,
         HostPaneRuntime, KeyHandling, LayoutControlEvent, MultiPaneTui, PaneMouseProtocol,
         PaneTextSelection, PaneViewState, PendingEscape, RemoteInput, RemoteSubscriptionState,
-        ScreenCell, ShareCopy, ShareView, SharedLayoutRuntime, SharedLocalPane, UiIntent,
-        grid_for_pane, initial_root_pane_grid, is_chord_command, is_chord_navigation,
-        lease_allows_held_input, member_label, pane_border_color, pane_title, pane_wire_id,
-        reconcile_remote_control_attempt, remote_input_decision, render_multi_pane,
-        render_multi_pane_with_copy_feedback, render_shared_multi_pane, selection_text, text_width,
-        viewed_screen, visible_leaf_panes,
+        ScreenCell, ShareCopy, SharedLayoutRuntime, SharedLocalPane, UiIntent, grid_for_pane,
+        initial_root_pane_grid, is_chord_command, is_chord_navigation, lease_allows_held_input,
+        member_label, pane_wire_id, reconcile_remote_control_attempt, remote_input_decision,
+        render_multi_pane, selection_text, viewed_screen,
     };
 
     #[test]
@@ -6133,51 +5754,6 @@ mod tests {
     }
 
     #[test]
-    fn shared_footer_places_rejection_notice_after_help() {
-        let snapshot = layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Leaf { pane_id: 1 },
-
-                title: None,
-            }],
-            &[(1, 1, 1)],
-        );
-        let tui = MultiPaneTui::new(snapshot).expect("layout");
-        let mut terminal = Terminal::new(TestBackend::new(160, 5)).expect("terminal");
-        terminal
-            .draw(|frame| {
-                render_shared_multi_pane(
-                    frame,
-                    &tui,
-                    &BTreeMap::new(),
-                    "",
-                    None,
-                    Some("layout request 5 rejected"),
-                    ShareView::default(),
-                    None,
-                );
-            })
-            .expect("draw");
-        let footer = (0..160)
-            .map(|x| terminal.backend().buffer()[(x, 4)].symbol())
-            .collect::<String>();
-        let help = footer.find("QUIT").expect("help rendered");
-        let notice = footer
-            .find("layout request 5 rejected")
-            .expect("rejection notice rendered");
-        assert!(help < notice, "notice sits after helper text");
-        assert_eq!(
-            terminal.backend().buffer()[(notice as u16, 4)].fg,
-            UiTheme::default().footer_orange
-        );
-        assert!(
-            !footer.trim_start().starts_with("layout request"),
-            "rejection no longer occupies the left status slot: {footer}"
-        );
-    }
-
-    #[test]
     fn pane_view_activity_transition_reports_a_redraw() {
         let snapshot = layout(
             vec![Tab {
@@ -6204,142 +5780,6 @@ mod tests {
                 ..tui.pane_view(1).expect("pane view").clone()
             },
         ));
-    }
-
-    #[test]
-    fn shared_renderer_keeps_join_commands_out_of_the_footer() {
-        let snapshot = layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Leaf { pane_id: 1 },
-
-                title: None,
-            }],
-            &[(1, 1, 1)],
-        );
-        let tui = MultiPaneTui::new(snapshot).expect("layout");
-        let mut terminal = Terminal::new(TestBackend::new(160, 5)).expect("terminal");
-        terminal
-            .draw(|frame| {
-                render_shared_multi_pane(
-                    frame,
-                    &tui,
-                    &BTreeMap::new(),
-                    "waiting for current pane reservation",
-                    None,
-                    None,
-                    ShareView {
-                        code: Some("TESTCODE"),
-                        ticket: Some("p2pmux-v1:TICKET"),
-                        notice: None,
-                    },
-                    None,
-                );
-            })
-            .expect("draw");
-        let footer = (0..160)
-            .map(|x| terminal.backend().buffer()[(x, 4)].symbol())
-            .collect::<String>();
-        assert!(footer.contains("waiting for current pane reservation"));
-        // Invite material lives behind Ctrl+S; a same-Mac-only command in the corner read as
-        // something a peer could run, and it never was.
-        assert!(!footer.contains("p2pmux join"));
-        assert!(!footer.contains("TESTCODE"));
-        assert!(footer.contains("SHARE"));
-    }
-
-    #[test]
-    fn share_modal_shows_the_ticket_and_the_code_with_their_reach() {
-        let snapshot = layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Leaf { pane_id: 1 },
-                title: None,
-            }],
-            &[(1, 1, 1)],
-        );
-        let mut tui = MultiPaneTui::new(snapshot).expect("layout");
-        tui.modal = super::ModalState::Share;
-        let mut terminal = Terminal::new(TestBackend::new(160, 24)).expect("terminal");
-        terminal
-            .draw(|frame| {
-                render_multi_pane_with_copy_feedback(
-                    frame,
-                    &tui,
-                    &BTreeMap::new(),
-                    None,
-                    None,
-                    ShareView {
-                        code: Some("TESTCODE"),
-                        ticket: Some("p2pmux-v1:TICKETVALUE"),
-                        notice: Some("✓ copied ticket"),
-                    },
-                    None,
-                    None,
-                );
-            })
-            .expect("draw");
-        let rendered = terminal
-            .backend()
-            .buffer()
-            .content
-            .iter()
-            .map(|cell| cell.symbol())
-            .collect::<String>();
-        assert!(rendered.contains("Share this session"));
-        assert!(rendered.contains("TICKET — works from any Mac"));
-        assert!(rendered.contains("p2pmux-v1:TICKETVALUE"));
-        assert!(rendered.contains("CODE — this Mac only"));
-        assert!(rendered.contains("TESTCODE"));
-        assert!(rendered.contains("✓ copied ticket"));
-        assert!(rendered.contains("COPY TICKET"));
-    }
-
-    #[test]
-    fn shared_footer_places_copy_feedback_after_quit_with_a_red_line_count() {
-        let snapshot = layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Leaf { pane_id: 1 },
-
-                title: None,
-            }],
-            &[(1, 1, 1)],
-        );
-        let tui = MultiPaneTui::new(snapshot).expect("layout");
-        let mut terminal = Terminal::new(TestBackend::new(120, 5)).expect("terminal");
-        terminal
-            .draw(|frame| {
-                render_shared_multi_pane(
-                    frame,
-                    &tui,
-                    &BTreeMap::new(),
-                    "",
-                    Some(3),
-                    None,
-                    ShareView::default(),
-                    None,
-                );
-            })
-            .expect("draw");
-        let footer = (0..120)
-            .map(|x| terminal.backend().buffer()[(x, 4)].symbol())
-            .collect::<String>();
-        let quit = footer.find("> QUIT").expect("quit help rendered");
-        let copied = footer
-            .find("copied 3 lines")
-            .expect("copy feedback rendered");
-        let count = footer.find("3 lines").expect("line count rendered");
-
-        assert!(copied > quit + "> QUIT".len());
-        assert_eq!(
-            terminal.backend().buffer()[(text_width(&footer[..copied]), 4)].fg,
-            Color::White
-        );
-        assert_eq!(
-            terminal.backend().buffer()[(text_width(&footer[..count]), 4)].fg,
-            Color::Rgb(255, 69, 0)
-        );
     }
 
     #[test]
@@ -7077,152 +6517,6 @@ mod tests {
     }
 
     #[test]
-    fn pane_title_uses_stable_leaf_order_and_control_state() {
-        let snapshot = layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Split {
-                    axis: Axis::LeftRight,
-                    first_share_bps: crate::layout::DEFAULT_FIRST_SHARE_BPS,
-                    first: Box::new(Node::Leaf { pane_id: 8 }),
-                    second: Box::new(Node::Leaf { pane_id: 3 }),
-                },
-
-                title: None,
-            }],
-            &[(3, 1, 1), (8, 1, 1)],
-        );
-        let mut members = snapshot.members.clone();
-        members[0].display_name = String::from("Host");
-        members.push(crate::layout::Member {
-            peer_id: b"guest".to_vec(),
-            endpoint_addr: b"guest-endpoint".to_vec(),
-            display_name: String::from("Guest"),
-        });
-
-        assert_eq!(visible_leaf_panes(&snapshot.tabs[0].root), vec![8, 3]);
-        assert_eq!(
-            pane_title(None, 1, b"host", Some(b""), false, false, &members, 80).to_string(),
-            "Pane #1 host: Host control: free"
-        );
-        assert_eq!(
-            pane_title(None, 2, b"host", Some(b"guest"), false, false, &members, 80).to_string(),
-            "Pane #2 host: Host control: Guest"
-        );
-        assert_eq!(
-            pane_title(None, 2, b"host", None, false, false, &members, 80).to_string(),
-            "Pane #2 host: Host control: …"
-        );
-        assert_eq!(
-            pane_title(None, 2, b"host", Some(b"guest"), true, false, &members, 80).to_string(),
-            "Pane #2 host: Host control: host-only"
-        );
-        assert_eq!(
-            pane_title(None, 2, b"host", Some(b"guest"), true, true, &members, 80).to_string(),
-            "Pane #2 host: Host control: exited"
-        );
-    }
-
-    #[test]
-    fn title_chrome_uses_custom_labels_and_cell_width_ellipsis() {
-        assert_eq!(
-            super::tab_label(Some("build logs"), 1, false, false),
-            "build logs"
-        );
-        assert_eq!(super::tab_label(None, 2, false, false), "Tab #2");
-        assert_eq!(
-            super::tab_label(Some("build logs"), 1, false, true),
-            "* build logs"
-        );
-        assert_eq!(super::truncate_trailing("界界", 3), "界…");
-        assert_eq!(super::truncate_trailing("界", 1), "…");
-        assert_eq!(super::truncate_trailing("title", 0), "");
-
-        let members = vec![crate::layout::Member {
-            peer_id: b"host".to_vec(),
-            endpoint_addr: b"endpoint".to_vec(),
-            display_name: String::from("Host"),
-        }];
-        assert_eq!(
-            pane_title(
-                Some("build logs"),
-                1,
-                b"host",
-                Some(b""),
-                false,
-                false,
-                &members,
-                12
-            )
-            .to_string(),
-            "build logs …"
-        );
-    }
-
-    #[test]
-    fn unread_agent_badges_use_the_same_starred_tab_label_for_hit_testing() {
-        let mut tui = MultiPaneTui::new(layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Split {
-                    axis: Axis::LeftRight,
-                    first_share_bps: crate::layout::DEFAULT_FIRST_SHARE_BPS,
-                    first: Box::new(Node::Leaf { pane_id: 1 }),
-                    second: Box::new(Node::Leaf { pane_id: 2 }),
-                },
-                title: Some(String::from("build")),
-            }],
-            &[(1, 2, 8), (2, 2, 8)],
-        ))
-        .expect("valid layout");
-        tui.unread_agent_panes.insert(2);
-        let area = Rect::new(0, 0, 80, 8);
-        let geometry = tui.geometry(area);
-        assert_eq!(
-            geometry.tab_labels[&1].width,
-            text_width(&super::tab_label(Some("build"), 1, true, true))
-        );
-
-        let mut terminal = Terminal::new(TestBackend::new(80, 8)).expect("test terminal");
-        terminal
-            .draw(|frame| render_multi_pane(frame, &tui, &BTreeMap::new()))
-            .expect("render");
-        let tab_bar = (0..80)
-            .map(|x| terminal.backend().buffer()[(x, 0)].symbol())
-            .collect::<String>();
-        let pane_titles = (0..80)
-            .map(|x| terminal.backend().buffer()[(x, 1)].symbol())
-            .collect::<String>();
-        assert!(tab_bar.contains("* build"));
-        assert!(pane_titles.contains("* Pane #2"));
-    }
-
-    #[test]
-    fn locked_badge_keeps_right_title_space_on_narrow_chrome() {
-        let mut snapshot = layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Leaf { pane_id: 1 },
-                title: None,
-            }],
-            &[(1, 2, 2)],
-        );
-        snapshot.panes.get_mut(&1).expect("pane").locked = true;
-        snapshot.members[0].display_name = String::from("Host");
-        let tui = MultiPaneTui::new(snapshot).expect("layout");
-        let mut terminal = Terminal::new(TestBackend::new(18, 4)).expect("terminal");
-
-        terminal
-            .draw(|frame| render_multi_pane(frame, &tui, &BTreeMap::new()))
-            .expect("draw");
-        let chrome = (0..18)
-            .map(|x| terminal.backend().buffer()[(x, 1)].symbol())
-            .collect::<String>();
-        assert!(chrome.contains("locked by Host"), "{chrome}");
-        assert!(!chrome.contains("Pane #1"));
-    }
-
-    #[test]
     fn rename_prompt_captures_chord_target_edits_and_consumes_modal_input() {
         let area = Rect::new(0, 0, 80, 24);
         let mut tui = MultiPaneTui::new(layout(
@@ -7403,124 +6697,6 @@ mod tests {
             KeyHandling::Consumed(vec![UiIntent::DeleteTab { tab_id: 1 }])
         );
         assert!(!tui.modal_open());
-    }
-
-    #[test]
-    fn free_panes_use_a_mid_gray_border_when_hovered_unfocused() {
-        let theme = UiTheme::default();
-        assert_eq!(
-            pane_border_color(
-                &theme,
-                false,
-                Some(b""),
-                false,
-                true,
-                false,
-                ChordMode::None
-            ),
-            Color::White
-        );
-        assert_eq!(
-            pane_border_color(
-                &theme,
-                false,
-                Some(b""),
-                false,
-                false,
-                true,
-                ChordMode::None
-            ),
-            Color::Gray
-        );
-        assert_eq!(
-            pane_border_color(
-                &theme,
-                false,
-                Some(b""),
-                false,
-                false,
-                false,
-                ChordMode::None
-            ),
-            Color::DarkGray
-        );
-        assert_eq!(
-            pane_border_color(&theme, false, None, false, true, false, ChordMode::None),
-            Color::Yellow
-        );
-        assert_eq!(
-            pane_border_color(&theme, false, None, false, false, true, ChordMode::None),
-            Color::Gray
-        );
-        assert_eq!(
-            pane_border_color(&theme, false, None, false, false, false, ChordMode::None),
-            Color::DarkGray
-        );
-        assert_eq!(
-            pane_border_color(
-                &theme,
-                false,
-                Some(b"guest"),
-                true,
-                true,
-                true,
-                ChordMode::None
-            ),
-            Color::Rgb(255, 69, 0)
-        );
-        assert_eq!(
-            pane_border_color(
-                &theme,
-                false,
-                Some(b"guest"),
-                false,
-                false,
-                true,
-                ChordMode::None
-            ),
-            Color::Rgb(255, 69, 0)
-        );
-
-        let themed = UiTheme {
-            pane_border_remote_control: Color::Rgb(1, 2, 3),
-            ..Default::default()
-        };
-        assert_eq!(
-            pane_border_color(
-                &themed,
-                false,
-                Some(b"guest"),
-                false,
-                false,
-                false,
-                ChordMode::None,
-            ),
-            Color::Rgb(1, 2, 3)
-        );
-        assert_eq!(
-            pane_border_color(
-                &theme,
-                false,
-                Some(b"guest"),
-                true,
-                true,
-                true,
-                ChordMode::Pane,
-            ),
-            theme.pane_border_chord_focused
-        );
-        assert_eq!(
-            pane_border_color(
-                &theme,
-                false,
-                Some(b"guest"),
-                true,
-                true,
-                true,
-                ChordMode::Tab,
-            ),
-            theme.pane_border_remote_control
-        );
     }
 
     #[test]
@@ -7757,197 +6933,6 @@ mod tests {
 
         assert_eq!(tui.current_tab(), 1);
         assert_eq!(tui.focused_pane(), 1);
-    }
-
-    #[test]
-    fn chrome_reports_the_pane_host_and_controller() {
-        let mut tui = MultiPaneTui::new(layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Leaf { pane_id: 1 },
-
-                title: None,
-            }],
-            &[(1, 2, 2)],
-        ))
-        .expect("valid layout");
-        tui.set_pane_view(
-            1,
-            PaneViewState {
-                ready: true,
-                controller_peer_id: Some(b"peer".to_vec()),
-                controller_active: true,
-                scrollback: 0,
-            },
-        );
-        let mut terminal = Terminal::new(TestBackend::new(36, 6)).expect("test terminal");
-
-        terminal
-            .draw(|frame| render_multi_pane(frame, &tui, &BTreeMap::new()))
-            .expect("render");
-        let buffer = terminal.backend().buffer();
-
-        assert_eq!(buffer[(0, 1)].symbol(), "┌");
-        assert_eq!(buffer[(1, 1)].symbol(), " ");
-        assert_eq!(buffer[(2, 1)].symbol(), "P");
-        assert!((2..9).all(|x| buffer[(x, 1)].modifier.contains(Modifier::BOLD)));
-        assert!(!buffer[(9, 1)].modifier.contains(Modifier::BOLD));
-        assert!(buffer.content.iter().any(|cell| cell.symbol() == "h"));
-        assert!(buffer.content.iter().any(|cell| cell.symbol() == "t"));
-    }
-
-    #[test]
-    fn pane_badge_uses_the_snapshot_host_not_mutable_view_state() {
-        let mut tui = MultiPaneTui::new(layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Leaf { pane_id: 1 },
-
-                title: None,
-            }],
-            &[(1, 2, 2)],
-        ))
-        .expect("valid layout");
-        tui.set_pane_view(
-            1,
-            PaneViewState {
-                ready: true,
-                controller_peer_id: None,
-                controller_active: false,
-                scrollback: 0,
-            },
-        );
-        let mut terminal = Terminal::new(TestBackend::new(40, 6)).expect("test terminal");
-
-        terminal
-            .draw(|frame| render_multi_pane(frame, &tui, &BTreeMap::new()))
-            .expect("render");
-        let title = (0..40)
-            .map(|x| terminal.backend().buffer()[(x, 1)].symbol())
-            .collect::<String>();
-
-        assert!(title.contains("host: 686f7374"));
-        assert!(!title.contains("host: 66616b65"));
-    }
-
-    #[test]
-    fn focused_ready_pane_maps_its_visible_vt_cursor_into_the_letterboxed_viewport() {
-        let mut parser = vt100::Parser::new(1, 3, 0);
-        parser.process(b"ab");
-        let mut tui = MultiPaneTui::new(layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Leaf { pane_id: 1 },
-
-                title: None,
-            }],
-            &[(1, 1, 3)],
-        ))
-        .expect("valid layout");
-        tui.set_pane_view(
-            1,
-            PaneViewState {
-                ready: true,
-                controller_peer_id: None,
-                controller_active: false,
-                scrollback: 0,
-            },
-        );
-        let screens = BTreeMap::from([(1, parser.screen())]);
-        let mut terminal = Terminal::new(TestBackend::new(9, 7)).expect("test terminal");
-
-        terminal
-            .draw(|frame| render_multi_pane(frame, &tui, &screens))
-            .expect("render");
-
-        terminal.backend_mut().assert_cursor_position((3, 2));
-    }
-
-    #[test]
-    fn focused_pane_never_draws_a_hidden_or_clipped_vt_cursor() {
-        for sequence in [b"\x1b[?25l".as_slice(), b"abcd".as_slice()] {
-            let mut parser = vt100::Parser::new(1, 5, 0);
-            parser.process(sequence);
-            let mut tui = MultiPaneTui::new(layout(
-                vec![Tab {
-                    tab_id: 1,
-                    root: Node::Leaf { pane_id: 1 },
-
-                    title: None,
-                }],
-                &[(1, 1, 3)],
-            ))
-            .expect("valid layout");
-            tui.set_pane_view(
-                1,
-                PaneViewState {
-                    ready: true,
-                    controller_peer_id: None,
-                    controller_active: false,
-                    scrollback: 0,
-                },
-            );
-            let screens = BTreeMap::from([(1, parser.screen())]);
-            let mut terminal = Terminal::new(TestBackend::new(9, 7)).expect("test terminal");
-
-            terminal
-                .draw(|frame| render_multi_pane(frame, &tui, &screens))
-                .expect("render");
-
-            assert!(!terminal.backend().cursor_visible());
-        }
-    }
-
-    #[test]
-    fn fixed_grid_view_is_top_left_and_clears_letterbox_margins() {
-        let mut previous_parser = vt100::Parser::new(3, 6, 0);
-        previous_parser.process(b"XXXXXX\r\nYYYYYY\r\nZZZZZZ");
-        let previous_tui = MultiPaneTui::new(layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Leaf { pane_id: 1 },
-
-                title: None,
-            }],
-            &[(1, 3, 6)],
-        ))
-        .expect("valid layout");
-        let mut terminal = Terminal::new(TestBackend::new(8, 7)).expect("test terminal");
-        terminal
-            .draw(|frame| {
-                render_multi_pane(
-                    frame,
-                    &previous_tui,
-                    &BTreeMap::from([(1, previous_parser.screen())]),
-                )
-            })
-            .expect("initial render");
-
-        let mut parser = vt100::Parser::new(1, 5, 0);
-        parser.process(b"abcde");
-        let tui = MultiPaneTui::new(layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Leaf { pane_id: 1 },
-
-                title: None,
-            }],
-            &[(1, 1, 5)],
-        ))
-        .expect("valid layout");
-        let screens = BTreeMap::from([(1, parser.screen())]);
-
-        terminal
-            .draw(|frame| render_multi_pane(frame, &tui, &screens))
-            .expect("render");
-        let buffer = terminal.backend().buffer();
-
-        assert_eq!(buffer[(1, 2)].symbol(), "a");
-        assert_eq!(buffer[(5, 2)].symbol(), "e");
-        assert_eq!(buffer[(7, 3)].symbol(), "│");
-        for (x, y) in [(6, 2), (1, 3), (6, 3), (1, 4), (6, 4)] {
-            assert_eq!(buffer[(x, y)].symbol(), " ", "letterbox cell ({x}, {y})");
-        }
     }
 
     #[test]
@@ -8295,135 +7280,6 @@ mod tests {
         tui.exit_chord_mode();
 
         assert_eq!(tui.chord_mode(), ChordMode::None);
-    }
-
-    #[test]
-    fn chord_mode_chrome_switches_and_restores_remote_focused_pane_colors() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        tui.set_pane_view(
-            1,
-            PaneViewState::from_chrome(true, Some(b"guest".to_vec()), true),
-        );
-        let mut terminal = Terminal::new(TestBackend::new(80, 8)).expect("test terminal");
-
-        tui.chord_mode = ChordMode::Pane;
-        terminal
-            .draw(|frame| render_multi_pane(frame, &tui, &BTreeMap::new()))
-            .expect("render");
-        let buffer = terminal.backend().buffer();
-        assert_eq!(buffer[(0, 1)].fg, tui.theme.pane_border_chord_focused);
-        assert_eq!(
-            (0..9).map(|x| buffer[(x, 7)].symbol()).collect::<String>(),
-            "PANE MODE"
-        );
-
-        tui.chord_mode = ChordMode::Tab;
-        terminal
-            .draw(|frame| render_multi_pane(frame, &tui, &BTreeMap::new()))
-            .expect("render");
-        let buffer = terminal.backend().buffer();
-        assert_eq!(buffer[(0, 1)].fg, tui.theme.pane_border_remote_control);
-        assert_eq!(
-            (0..8).map(|x| buffer[(x, 7)].symbol()).collect::<String>(),
-            "TAB MODE"
-        );
-
-        tui.chord_mode = ChordMode::None;
-        terminal
-            .draw(|frame| render_multi_pane(frame, &tui, &BTreeMap::new()))
-            .expect("render");
-        let buffer = terminal.backend().buffer();
-        assert_eq!(buffer[(0, 1)].fg, tui.theme.pane_border_remote_control);
-        assert!(
-            (0..4)
-                .map(|x| buffer[(x, 7)].symbol())
-                .collect::<String>()
-                .starts_with("Ctrl")
-        );
-    }
-
-    #[test]
-    fn themed_tui_overrides_footer_chrome() {
-        let theme = UiTheme {
-            footer_background: Color::Rgb(1, 2, 3),
-            footer_orange: Color::Rgb(4, 5, 6),
-            ..Default::default()
-        };
-        let tui = MultiPaneTui::with_theme(
-            layout(
-                vec![Tab {
-                    tab_id: 1,
-                    root: Node::Leaf { pane_id: 1 },
-                    title: None,
-                }],
-                &[(1, 2, 2)],
-            ),
-            theme,
-        )
-        .expect("valid layout");
-        let mut terminal = Terminal::new(TestBackend::new(120, 4)).expect("test terminal");
-
-        terminal
-            .draw(|frame| render_multi_pane(frame, &tui, &BTreeMap::new()))
-            .expect("render");
-        let footer = terminal.backend().buffer();
-        assert_eq!(footer[(0, 3)].bg, theme.footer_background);
-        assert_eq!(footer[(0, 3)].fg, theme.footer_orange);
-    }
-
-    #[test]
-    fn tab_bar_uses_a_branded_footer_like_strip_and_highlights_the_active_tab() {
-        let mut tui = MultiPaneTui::new(layout(
-            vec![
-                Tab {
-                    tab_id: 10,
-                    root: Node::Leaf { pane_id: 1 },
-
-                    title: None,
-                },
-                Tab {
-                    tab_id: 20,
-                    root: Node::Leaf { pane_id: 2 },
-
-                    title: None,
-                },
-            ],
-            &[(1, 2, 2), (2, 2, 2)],
-        ))
-        .expect("valid layout");
-        let mut terminal = Terminal::new(TestBackend::new(32, 4)).expect("test terminal");
-
-        terminal
-            .draw(|frame| render_multi_pane(frame, &tui, &BTreeMap::new()))
-            .expect("render");
-        let buffer = terminal.backend().buffer();
-        let tab_bar = (0..32).map(|x| buffer[(x, 0)].symbol()).collect::<String>();
-
-        assert!(tab_bar.starts_with("p2pmux │  Tab #1  · Tab #2"));
-        assert!(tab_bar.contains("Tab #1"));
-        assert!(tab_bar.contains("Tab #2"));
-        assert_eq!(buffer[(0, 0)].fg, Color::White);
-        assert_eq!(buffer[(0, 0)].bg, Color::Rgb(30, 30, 30));
-        assert_eq!(buffer[(9, 0)].fg, Color::White);
-        assert_eq!(buffer[(9, 0)].bg, Color::Rgb(220, 50, 47));
-        assert_eq!(buffer[(20, 0)].fg, Color::White);
-        assert_eq!(buffer[(20, 0)].bg, Color::Rgb(30, 30, 30));
-
-        tui.chord_mode = ChordMode::Tab;
-        terminal
-            .draw(|frame| render_multi_pane(frame, &tui, &BTreeMap::new()))
-            .expect("render");
-        let buffer = terminal.backend().buffer();
-        assert_eq!(buffer[(9, 0)].fg, Color::White);
-        assert_eq!(buffer[(9, 0)].bg, Color::Rgb(220, 50, 47));
-        assert_eq!(buffer[(20, 0)].fg, Color::DarkGray);
-        assert_eq!(buffer[(20, 0)].bg, Color::Rgb(30, 30, 30));
-
-        tui.chord_mode = ChordMode::None;
-        terminal
-            .draw(|frame| render_multi_pane(frame, &tui, &BTreeMap::new()))
-            .expect("render");
-        assert_eq!(terminal.backend().buffer()[(20, 0)].fg, Color::White);
     }
 
     #[test]
