+++ src/tui/render/mod.rs
@@ -3,4 +3,5 @@
 pub(in crate::tui) mod agents;
 pub(in crate::tui) mod footer;
 pub(in crate::tui) mod modals;
+pub(in crate::tui) mod panes;
 pub(in crate::tui) mod vt;
